fixed database issue

+++ database/migrations/2015_04_16_135740_create_pesanans_table.php
@@ -16,7 +16,7 @@ class CreatePesanansTable extends Migration {
 		{
 			$table->increments('id');
 			$table->integer('menu_id')->unsigned();
-			$table->integer('jumlah');
+			$table->decimal('jumlah');
 			$table->timestamp('tanggal');
 			$table->timestamps();
